% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniVRips.R
\name{aniVRips}
\alias{aniVRips}
\title{Create animated Vietoris-Rips complexes}
\usage{
aniVRips(
  df,
  framerate,
  extra_thresh = 0.25,
  inc_barcode = FALSE,
  filename = "VRComplexAni",
  title = "Animation of V-R Complex over Tau",
  description =
    "This animation represents the creation of the V-R complex of some set \\n\\n                                      of data. Very cool!"
)
}
\arguments{
\item{df}{A data frame with two columns to create the Vietoris-Rips from}

\item{framerate}{The number of frames you want the animation to use}

\item{extra_thresh}{The additional threshold for the function to use when graphing the complex. This adjusts the bounds of the graph.}

\item{inc_barcode}{If [TRUE], the animation include the persistence barcode plot.}

\item{filename}{The file name of the .html file.}

\item{title}{The title of the .html file.}

\item{description}{The description of the animation.}
}
\value{
An HTML file and its corresponding plots.
}
\description{
[aniVRips()] takes the input of a dataframe with two columns, 
calculates the persistence homology, and creates an HTML animation showing 
the formation of the complex over tau.
}
