% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Find the Euclidean distance between two points}
\usage{
distance(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1, x2, y2}{coordinate point values.}
}
\value{
a numeric distance
}
\description{
[distance()] is a function to calculate distance between two points.
This is a helper function for [aniVRips()] and [staticVRips()]
}
