% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\title{Find intersections of circles given a value of tau}
\usage{
intersection(df, tau)
}
\arguments{
\item{df}{A dataframe with two columns}

\item{tau}{A radial distance to calculate if the circles intersect.}
}
\value{
A dataframe of intersections given tau.
}
\description{
Find intersections of circles given a value of tau
}
