% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticVRips.R
\name{staticVRips}
\alias{staticVRips}
\title{Create static Vietoris-Rips complexes}
\usage{
staticVRips(
  df,
  tau,
  extra_thresh = 0.25,
  inc_barcode = FALSE,
  png = FALSE,
  filename = "VRComplexStatic_plot",
  ...
)
}
\arguments{
\item{df}{A data frame with two columns to create the Vietoris-Rips from.}

\item{tau}{A value of to compute the Vietoris-Rips for.}

\item{extra_thresh}{The additional threshold for the function to use when graphing the complex. This adjusts the bounds of the graph.}

\item{inc_barcode}{If [TRUE], the image includes the persistence barcode plot.}

\item{png}{If [TRUE], the png of the output is saved in the current directory.}

\item{filename}{The name of the png file.}

\item{...}{Addtional parameters that can be passed through [ggsave()]}
}
\value{
A plot of the Vietoris-Rips complex with given value tau and the barcode plot.

#' @examples
# ex_data2_x <- c(-2, 0, 0, 2)
ex_data2_y <- c(0, 1, -1, 0)
ex_data2 <- tibble(ex_data2_x, ex_data2_y)
colnames(ex_data2) <- c('Boop', 'Bop')

staticVRips(ex_data2, tau = 1.2, inc_barcode = TRUE)
}
\description{
[staticVRips()] takes the input of a dataframe with two columns, calculates the persistence homology,
}
