% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_calculate.R
\name{triangle_calculate}
\alias{triangle_calculate}
\title{Find triangles between three points}
\usage{
triangle_calculate(df, r)
}
\arguments{
\item{df}{A dataframe with two columns}

\item{r}{A radial distance to calculate if triangles are formed.}
}
\value{
A dataframe with three columns x, y, and group where group is the indicator
of what triangle the points are apart of.
}
\description{
[trianglecalculate()] is a function to calculate triangles between three points.
This is a helper function for [aniVRips()] and [staticVRips()]
}
